+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- added Grafana dashboard for device
 - added custom prometheus exporter
 
 ## v0.0.1